+++ src/icons.js
@@ -4,23 +4,14 @@
 const sharp = require('sharp');
 
 /**
- * Sizes for all app icons
+ * Internal dependencies
  */
-const iconSizes = [
-	'72x72',
-	'96x96',
-	'128x128',
-	'144x144',
-	'152x152',
-	'192x192',
-	'384x384',
-	'512x512'
-];
+const { iconSizes } = require('./sizes');
 
 /**
  * Generate all app icons
  */
-module.exports = function(icon, folder) {
+module.exports = (icon, folder) => {
 	for (const size of iconSizes) {
 		const [width, height] = size.split('x');
