+++ src/manifest.js
@@ -1,8 +1,18 @@
 /**
+ * Internal dependencies
+ */
+const { iconSizes } = require('./sizes');
+
+const icons = iconSizes.map(size => ({
+	src: `icons/icon-${size}.png`,
+	sizes: size,
+	type: 'image/png'
+}));
+
+/**
  * Generate a manifest.json file
  */
-module.exports = function(name) {
-	return `{
+module.exports = name => `{
 	"lang": "en",
 	"dir": "ltr",
 	"name": "${name}",
@@ -14,48 +24,6 @@ module.exports = function(name) {
 	"orientation": "portrait",
 	"scope": "/",
 	"start_url": "/",
-	"icons": [
-		{
-			"src": "icons/icon-72x72.png",
-			"sizes": "72x72",
-			"type": "image/png"
-		},
-		{
-			"src": "icons/icon-96x96.png",
-			"sizes": "96x96",
-			"type": "image/png"
-		},
-		{
-			"src": "icons/icon-128x128.png",
-			"sizes": "128x128",
-			"type": "image/png"
-		},
-		{
-			"src": "icons/icon-144x144.png",
-			"sizes": "144x144",
-			"type": "image/png"
-		},
-		{
-			"src": "icons/icon-152x152.png",
-			"sizes": "152x152",
-			"type": "image/png"
-		},
-		{
-			"src": "icons/icon-192x192.png",
-			"sizes": "192x192",
-			"type": "image/png"
-		},
-		{
-			"src": "icons/icon-384x384.png",
-			"sizes": "384x384",
-			"type": "image/png"
-		},
-		{
-			"src": "icons/icon-512x512.png",
-			"sizes": "512x512",
-			"type": "image/png"
-		}
-	],
+	"icons": ${JSON.stringify(icons, null, '\t')},
 	"splash_pages": null
 }`;
-};
